misc/HtmlTestSuite/results/test-1012.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Inertness with modal dialogs and iframes</title>
<link href="mailto:obrufau@igalia.com" rel="author" title="Oriol Brufau"/>
<link href="https://html.spec.whatwg.org/multipage/interaction.html#inert" rel="help"/>
<meta content="Checks that a modal dialog marks outer nodes as inert,&#xA;  but only in its document, not in the parent browsing context,&#xA;  nor in nested browsing contexts." name="assert"/>
</head><body><div id="log"/>
<div id="wrapper">
  (main document: outer text)
  <iframe id="outerIframe" srcdoc="&#xA;    &lt;div id='wrapper'&gt;&#xA;      (outer iframe: outer text)&#xA;      &lt;dialog id='dialog' style='display: block'&gt;&#xA;        (outer iframe: dialog)&#xA;      &lt;/dialog&gt;&#xA;    &lt;/div&gt;&#xA;  "/>
  <dialog id="dialog" style="display: block">
    (main document: dialog)
    <iframe id="innerIframe" srcdoc="&#xA;      &lt;div id='wrapper'&gt;&#xA;        (inner iframe: outer text)&#xA;        &lt;dialog id='dialog' style='display: block'&gt;&#xA;          (inner iframe: dialog)&#xA;        &lt;/dialog&gt;&#xA;      &lt;/div&gt;&#xA;    "/>
  </dialog>
</div>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
const innerIframeWindow = innerIframe.contentWindow;
const outerIframeWindow = outerIframe.contentWindow;
promise_setup(async () =&gt; {
  for (let global of [innerIframeWindow, outerIframeWindow]) {
    if (global.location.href === "about:blank" ||
        global.document.readyState !== "complete") {
      await new Promise(resolve =&gt; {
        global.frameElement.addEventListener("load", resolve, {once: true});
      });
    }
  }
});
add_completion_callback(() =&gt; {
  for (let global of [window, innerIframeWindow, outerIframeWindow]) {
    global.getSelection().removeAllRanges();
  }
});

function checkSelection(global, expectedText) {
  const selection = global.getSelection();
  selection.selectAllChildren(global.wrapper);

  // Remove whitespace between parentheses since it varies among browsers,
  // but that's not relevant to this test.
  const actualText = selection.toString().replace(/\)\s*\(/g, ")(").trim();
  assert_equals(actualText, expectedText);
}

function showModals(test, globals) {
  for (let global of globals) {
    global.dialog.showModal();
    test.add_cleanup(() =&gt; { global.dialog.close(); });
  }
}

promise_test(async function() {
  checkSelection(window, "(main document: outer text)(main document: dialog)");
  checkSelection(innerIframeWindow, "(inner iframe: outer text)(inner iframe: dialog)");
  checkSelection(outerIframeWindow, "(outer iframe: outer text)(outer iframe: dialog)");
}, "Initially, no node is inert");

promise_test(async function() {
  showModals(this, [outerIframeWindow]);

  checkSelection(window, "(main document: outer text)(main document: dialog)");
  checkSelection(innerIframeWindow, "(inner iframe: outer text)(inner iframe: dialog)");
  checkSelection(outerIframeWindow, "(outer iframe: dialog)");
}, "Modal dialog in the outer iframe marks outer nodes in that iframe as inert.");

promise_test(async function() {
  showModals(this, [innerIframeWindow]);

  checkSelection(window, "(main document: outer text)(main document: dialog)");
  checkSelection(innerIframeWindow, "(inner iframe: dialog)");
  checkSelection(outerIframeWindow, "(outer iframe: outer text)(outer iframe: dialog)");
}, "Modal dialog in the inner iframe marks outer nodes in that iframe as inert.");

promise_test(async function() {
  showModals(this, [innerIframeWindow, outerIframeWindow]);

  checkSelection(window, "(main document: outer text)(main document: dialog)");
  checkSelection(innerIframeWindow, "(inner iframe: dialog)");
  checkSelection(outerIframeWindow, "(outer iframe: dialog)");
}, "Modal dialogs in both iframes mark outer nodes in these iframes as inert.");

promise_test(async function() {
  showModals(this, [window]);

  checkSelection(window, "(main document: dialog)");
  checkSelection(innerIframeWindow, "(inner iframe: outer text)(inner iframe: dialog)");
  checkSelection(outerIframeWindow, "(outer iframe: outer text)(outer iframe: dialog)");
}, "Modal dialog in the main document marks outer nodes as inert. Contents of the outer iframe aren't marked as inert.");

promise_test(async function() {
  showModals(this, [innerIframeWindow, window]);

  checkSelection(window, "(main document: dialog)");
  checkSelection(innerIframeWindow, "(inner iframe: dialog)");
  checkSelection(outerIframeWindow, "(outer iframe: outer text)(outer iframe: dialog)");
}, "Modal dialogs in the main document and inner iframe mark outer nodes as inert. Contents of the outer iframe aren't marked as inert.");

promise_test(async function() {
  showModals(this, [outerIframeWindow, window]);

  checkSelection(window, "(main document: dialog)");
  checkSelection(innerIframeWindow, "(inner iframe: outer text)(inner iframe: dialog)");
  checkSelection(outerIframeWindow, "(outer iframe: dialog)");
}, "Modal dialogs in the main document and outer iframe mark outer nodes as inert. Contents of the outer iframe aren't marked as inert.");

promise_test(async function() {
  showModals(this, [innerIframeWindow, outerIframeWindow, window]);

  checkSelection(window, "(main document: dialog)");
  checkSelection(innerIframeWindow, "(inner iframe: dialog)");
  checkSelection(outerIframeWindow, "(outer iframe: dialog)");
}, "Modal dialogs in the main document and both iframes mark outer nodes as inert. Contents of the outer iframe aren't marked as inert.");
</script>
</body></html>
